chore: relocate cli auth route
